+++ data.tf
@@ -1,6 +1,6 @@
 data "aws_s3_bucket" "file_storage" {
   depends_on = [module.file_storage]
-  bucket     = local.bucket_name
+  bucket     = local.main_bucket_name
 }
 
 data "aws_sqs_queue" "file_storage" {