+++ examples/public-dns-external/variables.tf
@@ -33,7 +33,7 @@ variable "wandb_license" {
 variable "database_engine_version" {
   description = "Version for MySQL Auora"
   type        = string
-  default     = "8.0.mysql_aurora.3.02.2"
+  default     = "8.0.mysql_aurora.3.05.2"
 }
 
 variable "database_instance_class" {
@@ -127,7 +127,13 @@ variable "system_reserved_pid" {
 }
 
 variable "aws_loadbalancer_controller_tags" {
-  description = "(Optional) A map of AWS tags to apply to all resources managed by the load balancer controller"
+  description = "(Optional) A map of AWS tags to apply to all resources managed by load balancer and cluster"
   type        = map(string)
   default     = {}
+}
+
+variable "create_elasticache" {
+  type        = bool
+  default     = true
+  description = "whether to create an elasticache redis"
 }