+++ modules/app_eks/iam-policy-docs.tf
@@ -25,17 +25,46 @@ data "aws_iam_policy_document" "node_IMDSv2" {
   }
 }
 
+locals {
+  node_kms_actions = [
+    "kms:Encrypt",
+    "kms:Decrypt",
+    "kms:ReEncrypt*",
+    "kms:GenerateDataKey*",
+    "kms:DescribeKey"
+  ]
+}
 data "aws_iam_policy_document" "node_kms" {
   statement {
-    actions = [
-      "kms:Encrypt",
-      "kms:Decrypt",
-      "kms:ReEncrypt*",
-      "kms:GenerateDataKey*",
-      "kms:DescribeKey"
-    ]
+    actions   = local.node_kms_actions
     effect    = "Allow"
     resources = var.bucket_kms_key_arns
+  }
+  dynamic "statement" {
+    for_each = var.map_bucket_permissions.mode == "public" ? [1] : []
+    content {
+      actions   = local.node_kms_actions
+      effect    = "Allow"
+      resources = ["*"]
+      condition {
+        test     = "StringNotEquals"
+        variable = "kms:ResourceAccount"
+        values   = [data.aws_caller_identity.current.account_id]
+      }
+    }
+  }
+  dynamic "statement" {
+    for_each = var.map_bucket_permissions.mode == "restricted" ? [1] : []
+    content {
+      actions   = local.node_kms_actions
+      effect    = "Allow"
+      resources = ["*"]
+      condition {
+        test     = "StringEquals"
+        variable = "kms:ResourceAccount"
+        values   = var.map_bucket_permissions.accounts
+      }
+    }
   }
 }
 
@@ -55,6 +84,32 @@ data "aws_iam_policy_document" "node_s3" {
       "${var.bucket_arn}",
       "${var.bucket_arn}/*"
     ]
+  }
+  dynamic "statement" {
+    for_each = var.map_bucket_permissions.mode == "public" ? [1] : []
+    content {
+      actions   = ["s3:*"]
+      effect    = "Allow"
+      resources = ["*"]
+      condition {
+        test     = "StringNotEquals"
+        variable = "s3:ResourceAccount"
+        values   = [data.aws_caller_identity.current.account_id]
+      }
+    }
+  }
+  dynamic "statement" {
+    for_each = var.map_bucket_permissions.mode == "restricted" ? [1] : []
+    content {
+      actions   = ["s3:*"]
+      effect    = "Allow"
+      resources = ["*"]
+      condition {
+        test     = "StringEquals"
+        variable = "s3:ResourceAccount"
+        values   = var.map_bucket_permissions.accounts
+      }
+    }
   }
 }